+++ show/kdump.py
@@ -2,15 +2,13 @@ import click
 import utilities_common.cli as clicommon
 from swsssdk import ConfigDBConnector
 
-
 #
 # 'kdump command ("show kdump ...")
 #
-@click.group(cls=clicommon.AliasedGroup)
+@click.group(cls=clicommon.AliasedGroup, name="kdump")
 def kdump():
     """Show kdump configuration, status and information """
     pass
-
 
 @kdump.command('enabled')
 def enabled():
@@ -30,7 +28,6 @@ def enabled():
     else:
         click.echo("kdump is disabled")
 
-
 @kdump.command('status')
 def status():
     """Show kdump status"""
@@ -38,7 +35,6 @@ def status():
     clicommon.run_command("sonic-kdump-config --memory")
     clicommon.run_command("sonic-kdump-config --num_dumps")
     clicommon.run_command("sonic-kdump-config --files")
-
 
 @kdump.command('memory')
 def memory():
@@ -56,7 +52,6 @@ def memory():
                     kdump_memory = kdump_memory_from_db
     click.echo("Memory Reserved: {}".format(kdump_memory))
 
-
 @kdump.command('num_dumps')
 def num_dumps():
     """Show kdump max number of dump files"""
@@ -73,12 +68,10 @@ def num_dumps():
                     kdump_num_dumps = kdump_num_dumps_from_db
     click.echo("Maximum number of Kernel Core files Stored: {}".format(kdump_num_dumps))
 
-
 @kdump.command('files')
 def files():
     """Show kdump kernel core dump files"""
     clicommon.run_command("sonic-kdump-config --files")
-
 
 @kdump.command()
 @click.argument('record', required=True)
@@ -86,7 +79,6 @@ def files():
 def log(record, lines):
     """Show kdump kernel core dump file kernel log"""
     cmd = "sonic-kdump-config --file {}".format(record)
-
     if lines is not None:
         cmd += " --lines {}".format(lines)
